requirements.txt
astroid==2.2.5
Django==2.2.3
django-admin==1.3.2
django-excel-response2==2.0.8
django-six==1.0.4
djangorestframework==3.10.1
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
pkg-resources==0.0.0
pylint==2.3.1
pytz==2019.1
screen==1.0.1
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
wrapt==1.11.2
xlwt==1.3.0
